uso GLdouble y no GLfloat (Warnings)

+++ trunk/Practicas/Practica5/freeglut project 3D/freeglut project 3D/PV3D.cpp
@@ -86,9 +86,9 @@ GLdouble PV3D::magnitude()
 
 PV3D* PV3D::crossProduct(PV3D* v)
 {
-	GLfloat factori=(this->corY*v->getZ())-(this->corZ*v->getY());
-     GLfloat factorj=(this->corZ*v->getX())-(this->corX*v->getZ());
-     GLfloat factork=(this->corX*v->getY())-(this->corY*v->getX());
+	GLdouble factori=(this->corY*v->getZ())-(this->corZ*v->getY());
+     GLdouble factorj=(this->corZ*v->getX())-(this->corX*v->getZ());
+     GLdouble factork=(this->corX*v->getY())-(this->corY*v->getX());
 
      return new PV3D(factori,factorj,factork);
 
@@ -96,7 +96,7 @@ PV3D* PV3D::crossProduct(PV3D* v)
 
 PV3D* PV3D::matrixProduct(PV3D* Nt,PV3D* Bt,PV3D* Tt,PV3D* Ct)
 {
-        GLfloat matriz[4][4];
+        GLdouble matriz[4][4];
                 matriz[0][0]=Nt->getX();
                 matriz[1][0]=Nt->getY();
                 matriz[2][0]=Nt->getZ();
@@ -118,9 +118,9 @@ PV3D* PV3D::matrixProduct(PV3D* Nt,PV3D* Bt,PV3D* Tt,PV3D* Ct)
                 matriz[3][3]=1;
 
 
-        GLfloat valx=matriz[0][0]*corX+matriz[0][1]*corY+matriz[0][2]*corZ+matriz[0][3]*pv;
-        GLfloat valy=matriz[1][0]*corX+matriz[1][1]*corY+matriz[1][2]*corZ+matriz[1][3]*pv;
-        GLfloat valz=matriz[2][0]*corX+matriz[2][1]*corY+matriz[2][2]*corZ+matriz[2][3]*pv;
+        GLdouble valx=matriz[0][0]*corX+matriz[0][1]*corY+matriz[0][2]*corZ+matriz[0][3]*pv;
+        GLdouble valy=matriz[1][0]*corX+matriz[1][1]*corY+matriz[1][2]*corZ+matriz[1][3]*pv;
+        GLdouble valz=matriz[2][0]*corX+matriz[2][1]*corY+matriz[2][2]*corZ+matriz[2][3]*pv;
         return new PV3D(valx,valy,valz);
 }
